Removed compiler warnings.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/trunk/supertuxkart@2384 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/homing.cpp
@@ -140,8 +140,8 @@ void Homing::update(float dt)
 
         ideal_direction.setInterpolate3(actual_direction.normalized(), ideal_direction, dt);
         
-        const int current_xy_speed = sqrt( actual_direction.getX()*actual_direction.getX() +
-                                           actual_direction.getY()*actual_direction.getY());
+        const int current_xy_speed = (int)sqrt( actual_direction.getX()*actual_direction.getX() +
+                                                actual_direction.getY()*actual_direction.getY());
         
         m_body->setLinearVelocity( btVector3(ideal_direction.getX()*current_xy_speed,
                                              ideal_direction.getY()*current_xy_speed,
